some new tentative graph functions

+++ visualization.py
@@ -20,9 +20,14 @@ from tqdm import tqdm
 # choose False to recalculate the position of Nodes every iteration (which significantly slows down the process)
 '''
 def visualize_list(nodesDict_list, adjMatrix_list, iterations, model_name, pos_lock=True):
-        print("Generating graphs...")
+        
+        print("\n----------------------------------")
+        print("      Begin Visualization...")
+        print("----------------------------------")
+        
+        print("\nGenerating graphs...")
         # create directories and generate correct absolute path name
-        path_network, path_node_histogram, path_edge_histogram, path_animation = creat_dir(model_name + " (network)"), creat_dir(model_name + " (node-histogram)"), creat_dir(model_name + " (edge-histogram)"), creat_dir("animation")
+        path_network, path_node_histogram, path_edge_histogram, path_animation, path_evolution = creat_dir(model_name + " (network)"), creat_dir(model_name + " (node-histogram)"), creat_dir(model_name + " (edge-histogram)"), creat_dir("animation"), creat_dir("evolution")
         
         # cleans directory containing the network and histogram visualizations
         for model_path in [path_network, path_node_histogram, path_edge_histogram]:
@@ -31,27 +36,23 @@ def visualize_list(nodesDict_list, adjMatrix_list, iterations, model_name, pos_l
                                 os.remove(os.path.join(root, file))
         
         # create graphs
-        
         for i in tqdm(range(0, iterations + 1)):
                 visualization(nodesDict_list[i], adjMatrix_list[i], path_network, path_node_histogram, path_edge_histogram, i, pos_lock)
-        
+                
         # compile PNGs into gif (for both network and histogram)
         generate_gif(model_name + " (network)", path_network, path_animation)
-        generate_gif(model_name + " (edge-histogram)", path_edge_histogram, path_animation)
+        if generate_gif(model_name + " (edge-histogram)", path_edge_histogram, path_animation): print("No edge to plot!")
         generate_gif(model_name + " (node-histogram)", path_node_histogram, path_animation)
+        
+        # generating graph over time for node status + edge type
+        plot_status_over_time(nodesDict_list, model_name, path_evolution)
+        
+        dir_path = os.path.dirname(os.path.realpath(__file__))
+        print("\n----------------------------------")
+        print("     Visualization Complete!")
+        print("----------------------------------\n")
+        print("View Visualization in: " + dir_path + "/visualization" + "\n")
 
-
-# helper function: create directory if not already exists
-def creat_dir(folder_name):
-        dir_path = os.path.dirname(os.path.realpath(__file__))          # NOT os.getcwd() <——> this incantation is faulty
-        path = "{directory}/{subdirectory}/".format(directory = dir_path, subdirectory=folder_name)
-        mode = 0o755
-        # output graph png
-        try:  
-                os.makedirs(path, mode)
-        except OSError as error:  
-                pass
-        return path
 
 '''
 func generate_png_csv
@@ -74,12 +75,15 @@ func generate_png_csv
                 if both stubs connect to defecting nodes, edge = red
                 if one stub connects to cooperating node and the other defecting node, edge = yellow
 '''
+# node colors [0,1] : green, red (orangish), 
+# edge colors [2,3,4] : dark green, orange(leaning yellow), red (leaning magenta)
+color = ['#03b500', '#b52a00', "#077d11", "#ffbb3d", "#db0231"]    
 optimized_pos, position_lock = False, False            #setting position variables
+good_good_edges_list, mixed_edges_list, bad_bad_edges_list = [], [], []
+
 def visualization(nodes, adj, path_network, path_node_histogram, path_edge_histogram, index=-1, pos_lock = True, color_edges=True):
+        global color
         G = nx.convert_matrix.from_numpy_matrix(adj)
-        # node colors [0,1] : green, red (orangish), 
-        # edge colors [2,3,4] : dark green, yellow, red (leaning magenta)
-        color = ['#03b500', '#b52a00', "#005907", "#f5f122", "#db0231"]         
         nodes_color = [color[1] if (nodes[i].status == 1) else color[0] for i in range (len(nodes))]
 
         # node positioning
@@ -127,6 +131,13 @@ def visualization(nodes, adj, path_network, path_node_histogram, path_edge_histo
                 nx.draw_networkx_edges(G,optimized_pos,
                         edgelist=bad_bad_edges,
                         width=edge_width,alpha=edge_alpha,edge_color=color[4])
+                
+                # record edge type status for later plot
+                global good_good_edges_list, mixed_edges_list, bad_bad_edges_list
+                good_good_edges_list.append(good_good_edges)
+                bad_bad_edges_list.append(bad_bad_edges)
+                mixed_edges_list.append(mixed_edges)
+                
         
         # saving network graphs and histograms as PNGs in separate folders
         if index != -1: plt.savefig(path_network + "net-" + repr(index) + ".png", format="PNG")      # output graph png 
@@ -152,7 +163,11 @@ def visualization(nodes, adj, path_network, path_node_histogram, path_edge_histo
         plt.xlabel("Edge Type")
         
         plt.xticks(bar_spacing, edge_types)
-        plt.ylim([0, len(good_good_edges) + len(bad_bad_edges) + len(mixed_edges)])
+
+        top = len(good_good_edges) + len(bad_bad_edges) + len(mixed_edges)
+        if top <= 0: return 1            # since no more interactions, simulation is finished
+        
+        plt.ylim([0, top])
         plt.grid(True, axis='y')
         
         if index != -1: plt.savefig(path_edge_histogram + "edge-" + repr(index) + ".png", format="PNG")      # output graph png 
@@ -190,6 +205,86 @@ def visualization(nodes, adj, path_network, path_node_histogram, path_edge_histo
         return G        
 
 
+# graph the status of all nodes over time (proportion normalized by total)
+def plot_status_over_time(nodesDict_list, model_name, path_evolution):
+        print("\nPlotting node status evolution...")
+        
+        num_defect = []
+        num_nodes = len(nodesDict_list[0])
+        for nodeDict in nodesDict_list:
+                def_count = 0
+                for nodes in nodeDict.values():
+                        def_count += nodes.status         # assuming 1 represents defect, no multiedges, and nodes have binary status (0/1)
+                num_defect.append(def_count/num_nodes)
+        
+        num_coop = [1 - def_count for def_count in num_defect]      # populate cooperating array with the 'complement' of num_defect
+        
+        # convert to np array
+        iteration_count = [ i for i in range (0, len(nodesDict_list)) ]
+        x = np.array(iteration_count)
+        y1 = np.array(num_coop)    
+        y2 = np.array(num_defect)    
+        
+        # plot with line of best fit
+        plt.figure(figsize=(20, 15), dpi=80, facecolor='w', edgecolor='k')
+        plt.plot(x, y1, '-o', color=color[0], label="total cooperators")
+        plt.plot(x, y2, '-o', color=color[1], label="total defectors")
+
+        # plotting reference lines (.25, .5, .75)
+        plt.plot([-1, len(iteration_count)], [0.5, 0.5], 'k-', lw=1,dashes=[2, 2])
+        plt.plot([-1, len(iteration_count)], [0.25, 0.25], 'k-', lw=1,dashes=[2, 2])
+        plt.plot([-1, len(iteration_count)], [0.75, 0.75], 'k-', lw=1,dashes=[2, 2])
+        plt.ylim(0, 1)
+        
+        plt.xlabel('Iteration #')
+        plt.ylabel('Proportion')
+        plt.title("Proportion of Cooperators to Defectors over Iterations")
+        # plt.text(0, 1,'matplotlib', horizontalalignment='center',
+        #      verticalalignment='center'))
+        plt.legend()
+        plt.savefig(path_evolution + model_name + "-Evolution-Nodes.png", format="PNG")
+        plt.close()
+        
+        
+        ##### graph edge type evolution #####
+        
+        print("Plotting edge type evolution...\n")
+        # get edge counts
+        global good_good_edges_list, mixed_edges_list, bad_bad_edges_list
+        
+        total = [ len(good_good_edges_list[i]) + len(mixed_edges_list[i]) + len(bad_bad_edges_list[i]) 
+                 if len(good_good_edges_list[i]) + len(mixed_edges_list[i]) + len(bad_bad_edges_list[i]) > 0
+                 else -1
+                 for i in range (len(iteration_count)) ]
+        
+        good_edges_count = [ len(good_good_edges_list[i])/total[i] for i in range(len(iteration_count)) ]
+        mixed_edges_count = [ len(mixed_edges_list[i])/total[i] for i in range(len(iteration_count)) ]
+        bad_edge_count = [ len(bad_bad_edges_list[i])/total[i] for i in range(len(iteration_count)) ]
+
+        y1 = np.array(good_edges_count)    
+        y2 = np.array(mixed_edges_count)    
+        y3 = np.array(bad_edge_count)    
+        
+        # plot with line of best fit
+        plt.figure(figsize=(20, 15), dpi=80, facecolor='w', edgecolor='k')
+        plt.plot(x, y1, '-o', color=color[2], label="total good-good edges")
+        plt.plot(x, y2, '-o', color=color[3], label="total mixed edges")
+        plt.plot(x, y3, '-o', color=color[4], label="total bad-bad edges")
+
+        # plotting reference lines (.25, .5, .75)
+        plt.plot([-1, len(iteration_count)], [0.5, 0.5], 'k-', lw=1,dashes=[2, 2])
+        plt.plot([-1, len(iteration_count)], [0.25, 0.25], 'k-', lw=1,dashes=[2, 2])
+        plt.plot([-1, len(iteration_count)], [0.75, 0.75], 'k-', lw=1,dashes=[2, 2])
+        plt.ylim(0, 1)
+        
+        plt.xlabel('Iteration #')
+        plt.ylabel('Proportion')
+        plt.title("Proportion of Edge Types")
+
+        plt.legend()
+        plt.savefig(path_evolution + model_name + "--Evolution-Edges.png", format="PNG")
+        plt.close()
+
 '''
 func generate_gif
 @param 
@@ -200,7 +295,7 @@ func generate_gif
 Note: only call when all graphs have been generated using func visualization(...)
 '''
 def generate_gif(model_name, input_path, output_path):
-        print("Compiling GIF...")
+        print("\nCompiling GIF...")
         my_path = input_path
         # basically the python version of regular expression search of list segments
         # you can also use globbing and bash commands, but apparently wildcard symbols can be unstable across different shell versions
@@ -227,3 +322,15 @@ def generate_gif(model_name, input_path, output_path):
                                 pbar.update(os.stat(pic).st_size)
         print("GIF Generated: stored in {0}".format(output_path))
         
+
+# helper function: create directory if not already exists
+def creat_dir(folder_name, catagory="visualization"):
+        dir_path = os.path.dirname(os.path.realpath(__file__))          # NOT os.getcwd() <——> this incantation is faulty
+        path = "{directory}/{catagory}/{subdirectory}/".format(directory = dir_path, catagory=catagory, subdirectory=folder_name)
+        mode = 0o755
+        # output graph png
+        try:  
+                os.makedirs(path, mode)
+        except OSError as error:  
+                pass
+        return path
